tg_bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Vladimir Mashyanov <vova.mashyanov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.1"
bcrypt = "^4.2.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
aiosqlite = "^0.20.0"
mypy = "^1.13.0"
ruff = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
